Connit

+++ lesson_02/lesson_2_task_3.py
@@ -3,5 +3,5 @@ import math
 def square(x):
     return math.ceil(x * x)
 
-side_square = int(input("Введите сторону квадрата: "))
+side_square = float(input("Введите сторону квадрата: "))
 print(f"Площадь квадрата: {square(side_square)}")
